Refactored code!
Also trying to get the SubBranch (The one that is the child of another SubBranch) to rotate when the main Branch rotates. (I don't have to do this for SubBranches that are children of the main Branch, this is because for those... I don't reset the rotations, and it is also why for those I don't have to do trigonometry. Maybe one day I'll change this)