Refactored Code again!
Last time I think I just cleaned up code, but anyway... I managed to get the SubBranch which is a child of Another SubBranch to not reset the rotations. So now it works well without having to move around the angles in the background. This is my first bit of code to get it to work, you can see two squares (which were used to see where certain places in the rotated plane are)